requirements.txt
django>=2.0.8
django-filter==1.1.0
djangorestframework==3.7.7
psycopg2-binary==2.7.5
six==1.11.0
